Refactored test_cartesi_voucher.fish to handle cases where Docker is not running.

+++ backups/test_cartesi_voucher.fish
@@ -8,17 +8,25 @@ function test_cartesi_voucher
   or return
   echo "======== $_flag_path $_flag_p "
   set logfile $_flag_path"test.log"
-  echo "****logfile: $logfile" &|tee -a $logfile
+  echo "**** logfile: $logfile" &|tee -a $logfile
+  if not docker version >/dev/null
+    echo "docker isn't running :-("  &| tee -a $logfile
+    return
+  end
   docker image inspect coin-toss-contracts >/dev/null 2>&1; and docker image rm coin-toss-contracts; or true
   docker buildx bake -f docker-bake.hcl -f docker-bake.override.hcl --load
   fish -c "docker compose -f docker-compose.yml -f docker-compose.override.yml up"&
   while true
+    if not docker version >/dev/null
+      echo "docker isn't running :-( (in the while block loop)"  &| tee -a $logfile
+      return
+    end
     set hex_response (curl -X POST --data '{"jsonrpc":"2.0","method":"eth_blockNumber","params":[],"id":1}' $RPC_URL 2>/dev/null)
     # Check if the response is empty (server might not be running)
     if test -z "$hex_response"
-        echo "RPC server not available. Retrying in 10 seconds..."
-        sleep 10
-        continue
+      echo "RPC server not available. Retrying in 10 seconds..."
+      sleep 10
+      continue
     end
 
     # Process response
@@ -26,17 +34,42 @@ function test_cartesi_voucher
     set decimal_number (math $hex_number)
     echo "Current block number is $decimal_number."
 
-    set cut_off_block 28
+    set cut_off_block_load 28 # by this time, all should be loaded
 
     # Check if the block number is greater than 1
-    if test $decimal_number -gt $cut_off_block
+    if test $decimal_number -gt $cut_off_block_load
       echo "Block number is $decimal_number, which is greater than $cut_off_block."
       docker run --rm --net="host" ghcr.io/foundry-rs/foundry "cast send --private-key $PLAYER1_PRIVATE_KEY --rpc-url $RPC_URL $COIN_TOSS_ADDRESS \"set_dapp_address(address)\" $DAPP_ADDRESS"
       docker run --rm --net="host" ghcr.io/foundry-rs/foundry "cast send --private-key $PLAYER1_PRIVATE_KEY --rpc-url $RPC_URL $COIN_TOSS_ADDRESS \"play(address)\" $PLAYER2"
       docker run --rm --net="host" ghcr.io/foundry-rs/foundry "cast send --private-key $PLAYER2_PRIVATE_KEY --rpc-url $RPC_URL $COIN_TOSS_ADDRESS \"play(address)\" $PLAYER1"
       curl --data '{"id":1337,"jsonrpc":"2.0","method":"evm_increaseTime","params":[864010]}' http://localhost:8545
       curl --data '{"id":1337,"jsonrpc":"2.0","method":"evm_increaseTime","params":[864010]}' http://localhost:8545
+      break
+    end
 
+    sleep 10
+  end
+
+  while true
+    if not docker version >/dev/null
+      echo "docker isn't running :-( (in the while block loop)"  &| tee -a $logfile
+      return
+    end
+    set hex_response (curl -X POST --data '{"jsonrpc":"2.0","method":"eth_blockNumber","params":[],"id":1}' $RPC_URL 2>/dev/null)
+    # Check if the response is empty (server might not be running)
+    if test -z "$hex_response"
+      echo "front end RPC server not available. Retrying in 10 seconds..."
+      sleep 10
+      continue
+    end
+
+    # Process response
+    set hex_number (echo $hex_response | jq -r .result)
+    set decimal_number (math $hex_number)
+    echo "Current block number is $decimal_number."
+
+    set cut_off_block_interaction_wait 38 # by this time, all should be loaded
+    if test $decimal_number -gt $cut_off_block_interaction_wait
       cd ../rollups-examples/frontend-console/
       # yarn && yarn build
       yarn start notice list &| tee -a $logfile
@@ -52,15 +85,5 @@ function test_cartesi_voucher
 
     sleep 10
   end
+  echo "**** logfile: $logfile" &|tee -a $logfile
 end
-# test_cartesi_voucher
-
-# docker images && docker ps -a --no-trunc &&  docker volume ls && docker network ls
-# docker image rm coin-toss-contracts
-
-# docker stop $(docker ps -aq)
-# docker rm $(docker ps -aq) -f
-# docker rmi $(docker images -q) -f
-# docker volume rm $(docker volume ls -q) -f
-# # docker network ls | grep "bridge\|none\|host" | awk '/ / { print $1 }' | xargs -r docker network rm -f
-# docker system prune -a --volumes -f
